ReadOnlyInlineMixin was moved to utils.admin.py
