+++ api/utils/admin.py
@@ -1,0 +1,9 @@
+class ReadOnlyInlineMixin:
+    def has_change_permission(self, request, obj=None):
+        return False
+
+    def has_add_permission(self, request, obj):
+        return False
+
+    def has_delete_permission(self, request, obj=None):
+        return False
